+++ src/UiAdminBase/CHANGELOG.md
@@ -15,6 +15,10 @@ The version numbers are structured like `GENERATION.MAJOR.MINOR.PATCH`:
 
 ### Added
 
+- Add composer dependency `heptacom/heptaconnect-dataset-base: self.version`
+- Add composer dependency `heptacom/heptaconnect-portal-base: self.version`
+- Add UI admin action `\Heptacom\HeptaConnect\Ui\Admin\Base\Contract\Action\Portal\PortalEntityListUiActionInterface` with `\Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListCriteria` and `\Heptacom\HeptaConnect\Ui\Admin\Base\Action\Portal\PortalEntityList\PortalEntityListResult` to list supported entity types in a fresh portal node stack of the given criteria
+
 ### Changed
 
 ### Deprecated